# contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Jitong Server Contract
  version: 1.0.0
  description: Baseline operational contract. Business endpoints are added in later tracer bullets.
paths:
...
  /api/v1/users/search:
    get:
      operationId: searchUserByAccountNo
      summary: Search one discoverable user by exact account number
      security: [{ bearerAuth: [] }]
      parameters:
        - name: accountNo
          in: query
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]{10}$' }
      responses:
        '200':
          description: Minimal discoverable profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContactSearchResult' }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /api/v1/users/me/searchability:
    post:
      operationId: updateAccountSearchability
      summary: Enable or disable exact account-number discovery
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [searchable]
              properties:
                searchable: { type: boolean }
      responses:
        '204': { description: Searchability updated }
        '401':
          $ref: '#/components/responses/ApiError'
...
  /api/v1/contact-requests:
    get:
      operationId: listContactRequests
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Contact request history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContactRequestSummary' }
    post:
      operationId: createContactRequest
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateContactRequest' }
      responses:
        '200':
          description: Request state after creation or cross-request acceptance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContactRequestResponse' }
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
...
  /api/v1/contact-requests/{requestId}/accept:
    post:
      operationId: acceptContactRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
      responses:
        '200':
          description: Accepted request and shared conversation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContactRequestResponse' }
        '403':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /api/v1/contact-requests/{requestId}/reject:
    post:
      operationId: rejectContactRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
      responses:
        '200':
          description: Rejected request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContactRequestResponse' }
        '403':
          $ref: '#/components/responses/ApiError'
  /api/v1/contact-requests/{requestId}/cancel:
    post:
      operationId: cancelContactRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
      responses:
        '200':
          description: Cancelled request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContactRequestResponse' }
        '403':
          $ref: '#/components/responses/ApiError'
  /api/v1/contacts:
    get:
      operationId: listContacts
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Active contacts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContactSummary' }
  /api/v1/contacts/{userId}:
    delete:
      operationId: removeContact
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: Contact removed; history remains read-only }
        '404':
          $ref: '#/components/responses/ApiError'
  /api/v1/blocks/{userId}:
    post:
      operationId: blockUser
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: User blocked }
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: unblockUser
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: User unblocked; contact is not restored }
  /api/v1/conversations:
    get:
      operationId: listConversations
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: C2C conversations, including read-only history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConversationSummary' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  parameters:
    RequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ContactSearchResult:
      type: object
      additionalProperties: false
      required: [version, accountNo, displayName, avatarVersion, avatarFallback, relationship]
      properties:
        version: { const: 1 }
        accountNo: { type: string, pattern: '^[1-9][0-9]{10}$' }
        displayName: { type: string }
        avatarUrl: { type: [string, 'null'] }
        avatarVersion: { type: integer, format: int64, minimum: 0 }
        avatarFallback: { type: string, minLength: 1, maxLength: 2 }
        relationship:
          type: string
          enum: [NONE, PENDING_OUTGOING, PENDING_INCOMING, ACTIVE, REMOVED]
        pendingRequestId: { type: [string, 'null'], format: uuid }
    CreateContactRequest:
      type: object
      additionalProperties: false
      required: [accountNo]
      properties:
        accountNo: { type: string, pattern: '^[1-9][0-9]{10}$' }
        verification: { type: string, maxLength: 100 }
    ContactRequestResponse:
      type: object
      additionalProperties: false
      required: [version, requestId, requesterId, recipientId, status, verification, expiresAt]
      properties:
        version: { const: 1 }
        requestId: { type: string, format: uuid }
        requesterId: { type: string, format: uuid }
        recipientId: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, ACCEPTED, REJECTED, CANCELLED, EXPIRED] }
        verification: { type: string }
        expiresAt: { type: string, format: date-time }
        conversationId: { type: [string, 'null'], format: uuid }
    ContactRequestSummary:
      allOf:
        - $ref: '#/components/schemas/ContactRequestResponse'
        - type: object
          required: [incoming, peerAccountNo, peerDisplayName]
          properties:
            incoming: { type: boolean }
            peerAccountNo: { type: string, pattern: '^[1-9][0-9]{10}$' }
            peerDisplayName: { type: string }
    ContactSummary:
      type: object
      required: [version, userId, accountNo, displayName, conversationId, relationship, avatarVersion]
      properties:
        version: { const: 1 }
        userId: { type: string, format: uuid }
        accountNo: { type: string, pattern: '^[1-9][0-9]{10}$' }
        displayName: { type: string }
        conversationId: { type: string, format: uuid }
        relationship: { const: ACTIVE }
        avatarUrl: { type: [string, 'null'] }
        avatarVersion: { type: integer, format: int64, minimum: 0 }
    ConversationSummary:
      type: object
      required: [version, conversationId, peerUserId, peerAccountNo, peerDisplayName, status, relationship, blockedByMe, readSeq, peerReadSeq, avatarVersion, searchVisible, searchVisibleAfterSeq]
      properties:
        version: { const: 1 }
        conversationId: { type: string, format: uuid }
        peerUserId: { type: string, format: uuid }
        peerAccountNo: { type: [string, 'null'], pattern: '^[1-9][0-9]{10}$' }
        peerDisplayName: { type: string }
        status: { type: string, enum: [ACTIVE, READ_ONLY, DISSOLVED] }
        relationship: { type: string, enum: [ACTIVE, READ_ONLY] }
        blockedByMe: { type: boolean }
        readSeq: { type: integer, format: int64, minimum: 0 }
        peerReadSeq: { type: integer, format: int64, minimum: 0 }
        searchVisible: { type: boolean }
        searchVisibleAfterSeq: { type: integer, format: int64, minimum: 0 }
        avatarUrl: { type: [string, 'null'] }
        avatarVersion: { type: integer, format: int64, minimum: 0 }
...
  responses:
    ApiError:
      description: Versioned error response
      headers:
        X-Request-Id:
          description: UUIDv4 request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/json:
          schema:
            $ref: './schemas/error-response-v1.schema.json'